Meson default in ci (#1675)
* Switch CI to build using meson & ninja by default

* Add threadsafety option to meson

* Fix whitespace

* Don't set werror=true by default

* Use distro version of clang/llvm

* Drop buster and add bookworm docker test

* No point in adding languages to matrix

* Use clang-tools-15 for scan-build

* Fix tarbuild

* Disable werror when building Xen tools

* Add Ubuntu rolling container test

* disable-werror on QEMU during Xen build

* Build package using meson

* Use warning level 0 by default

* Don't build ovmf on bookworm

* Guard CLONE_NEWTIME

* Do -Wall -Wextra -Werror on debug build

* Add scripts/version.sh

* Remove package/patch-xen.sh

* Replace : with -

* Don't use artifact prefix

* Install vol3 using virtualenv

* Update delete old releases to 0.2.1

* Fix warning level and werror

* Use meson setup

* Guard variable used only for debug

+++ package/configure-xen.sh
@@ -2,13 +2,23 @@
 
 # Building OVMF is broken with gcc-11 (default on jammy)
 # Unfortunately we can't just set CC in the environment because the edk2 build system is insane
+# gcc-9 is not available on bookworm so there we build with no ovmf at all
 
+OVMF=""
 GCC_VERSION=$(gcc --version | head -n1 | awk '{ print $4 }' | awk -F'.' '{ print $1 }')
-if [ ${GCC_VERSION} -gt 9 ]; then
+if [ ${GCC_VERSION} -gt 9 ] && [ -f /usr/bin/gcc-9 ]; then
     rm /usr/bin/gcc
     ln -s /usr/bin/gcc-9 /usr/bin/gcc
+    OVMF="--enable-ovmf"
 fi
 
-./configure --prefix=/usr --enable-githttp --disable-pvshim --enable-systemd --enable-ovmf
+./configure --prefix=/usr --enable-githttp \
+    --disable-pvshim --disable-werror \
+    --with-extra-qemuu-configure-args="--disable-werror --disable-sdl" \
+    --enable-systemd $OVMF
+
+echo CONFIG_EXPERT=y > xen/.config
+echo CONFIG_MEM_SHARING=y >> xen/.config
+make -C xen olddefconfig
 
 exit 0
